added APIs

+++ littlelemon/restaurant/models.py
@@ -4,9 +4,9 @@ from django.db import models
 class Menu(models.Model):
     title = models.CharField(max_length=255)
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    inventory = models.SmallIntegerField()
+    inventory = models.IntegerField()
 
 class Booking(models.Model):
     name = models.CharField(max_length=255)
-    guests = models.SmallIntegerField()
+    guests = models.IntegerField()
     date = models.DateTimeField()
